pyproject.toml
[project]
name = "stockbot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["discord.py", "pydantic", "pydantic-settings", "yfinance", "aiohttp"]

[dependency-groups]
dev = ["ruff"]